Refactor Nostr channel hotspots into dedicated modules
